backend/requirements.txt
django
djangorestframework
psycopg2-binary
django-cors-headers
pypandoc-binary
python-docx
requests
djangorestframework-simplejwt
